// eat-n-repeat-frontend/eat-n-repeat-frontend/lib/admin/types.ts
export type Addon = {
  id: string;
  name: string;
  price: number;
  available: boolean;
  createdAt: string;
};

export type AddonInput = Omit<Addon, "id" | "createdAt">;